--- pages/stisla/dashboard/general.tsx
@@ -1,12 +1,0 @@
-import AdminLayout from "@layouts/admin-layout";
-import Head from "next/head";
-
-export default function DashboardGeneral() {
-  return (
-    <AdminLayout>
-      <Head>
-        <title> Dashboard Title </title>
-      </Head>
-    </AdminLayout>
-  );
-}
